Improve first-run workflow
The aim of this change is to reduce the chances of a user encountering
baffling errors that are difficult to diagnose. Since the Freckle API
does not return error messages, it can be hard to know what is wrong
when it doesn’t work.

One thing that could go wrong with the previous version was the username
being wrong. We avoid this possibility by authenticating the username
as part of the signup process, by prompting for username/password
rather than API key.